requirements.txt
pillow
requests
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
pyyaml
